refactor(errors): centralize custom errors
