+++ src/errors/index.tsx
@@ -1,7 +1,15 @@
 export {
   AgentCoreCLIError,
+  DeserializationError,
+  EmbeddedAssetNotFoundError,
   InputValidationError,
+  InvalidEnvironmentError,
+  NestedProjectError,
   NotImplementedError,
+  ProjectFileExistsError,
+  RuntimeInvokeInterruptedError,
+  RuntimeInvokeResponseError,
+  SourceResolutionError,
   type AgentCoreCLIErrorOptions,
 } from "./errors";
 export { ERROR_SOURCE } from "./types";
