+++ src/handlers/runtime/invoke/request.ts
@@ -1,8 +1,8 @@
 import { validateHeaderName, validateHeaderValue } from "node:http";
 import z from "zod";
 import type { GetAgentRuntimeResponse } from "@aws-sdk/client-bedrock-agentcore-control";
-import { InputValidationError } from "../../../errors";
-import { SourceResolutionError, SourceResolver } from "../../../io";
+import { InputValidationError, SourceResolutionError } from "../../../errors";
+import { SourceResolver } from "../../../io";
 import type { RuntimeInvokeRequest } from "../types";
 
 export const runtimeIdSchema = z
