+++ src/io/index.ts
@@ -1,4 +1,4 @@
 export { atomicWrite } from "./atomicWrite";
 export { FsReadWriteJson } from "./json";
-export { SourceResolutionError, SourceResolver, type SourceResolverConfig } from "./source";
+export { SourceResolver, type SourceResolverConfig } from "./source";
 export type { AppIO, ReadWriteJson } from "./types";
