+++ src/io/source.ts
@@ -1,7 +1,7 @@
 import { readFile } from "node:fs/promises";
 import { addAbortSignal } from "node:stream";
 import { buffer } from "node:stream/consumers";
-import { InputValidationError } from "../errors";
+import { SourceResolutionError } from "../errors";
 
 const FILE_PREFIX = "file://";
 const STDIN = "-";
@@ -10,13 +10,6 @@ export type SourceResolverConfig = {
   stdin: NodeJS.ReadStream;
   signal?: AbortSignal;
 };
-
-export class SourceResolutionError extends InputValidationError {
-  constructor(message: string, options?: ErrorOptions) {
-    super(message, options);
-    this.name = "SourceResolutionError";
-  }
-}
 
 export class SourceResolver {
   private stdinClaimedBy?: string;
