+++ src/tui/index.tsx
@@ -11,7 +11,7 @@ import {
 import type { AppIO } from "../io";
 import type { Core } from "../handlers/types";
 import { JsonKey } from "../handlers/keys";
-import { AgentCoreCLIError, ERROR_SOURCE, type AgentCoreCLIErrorOptions } from "../errors";
+import { InvalidEnvironmentError } from "../errors";
 
 // renderJson pretty-prints a value as indented JSON. It is the output
 // counterpart to renderTui: handlers call it to emit machine-readable results
@@ -74,11 +74,4 @@ export function renderTui(core: Core, io: AppIO): DefaultHandle {
 
     await renderTuiAt(ctx.require(PathKey), ctx, core, io);
   };
-}
-
-/** Error raised when detecting an invalid environment */
-export class InvalidEnvironmentError extends AgentCoreCLIError {
-  constructor(message?: string, options?: Omit<AgentCoreCLIErrorOptions, "source">) {
-    super(message, { ...options, source: ERROR_SOURCE.USER });
-  }
 }
